Created new socket file and destroyed socket utils file.
